+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Fixed
+- Upload applet `deprecated` and `published` versions are now correctly handled and show an info message
+
 ## [1.6.1] - 2024-09-30
 ### Fixed
 - Generate applet with `--language=typescript` will generate correct `webpack.config.js` file (with `fileExtensions` and `rules`)